feat: add UserAddedChainContext
