--- packages/extension-polkagate/src/fullscreen/addNewChain/types.tsx
@@ -1,9 +1,0 @@
-// Copyright 2019-2024 @polkadot/extension-polkagate authors & contributors
-// SPDX-License-Identifier: Apache-2.0
-
-export interface UserAddedEndpoint {
-  chain: string;
-  color: string;
-  endpoint: string;
-  priceId: string;
-}
